feat: 2-week sprint — causal fusion, online detection fix, walk-forward HPO
Sprint results (5 workstreams):

A) Causal fusion: equal-weight Berry+QFI+MLF achieves d=0.664 (12 crises),
   val d=0.560 — beats CUSUM val d=0.459 by 22% with no look-ahead.

B) Comprehensive metrics: Berry Phase Rate best overall rank (3.79/14 metrics).

C) 500-trial HPO confirms h=6, pca=8, rw=15, sphere, f01 are optimal for Berry.

D) Online detection AUC fixed from 0.50 to 0.844 via signal inversion
   (geometric features decrease during crises — manifold contraction).

E) Feature ablation: minimal features (d=0.468) beat enriched (d=0.304),
   random operators beat pca_inspired for Berry.

New experiments: walk_forward_hpo.py (nested Optuna HPO at each expanding
window — fully unbiased OOS), causal_fusion_final.py, online_detection_improved.py,
feature_engineering_ablation.py, comprehensive_metric_comparison.py, and more.

Paper updated: causal fusion section, online detection fix, abstract/intro/
conclusion rewritten. 37 pages, compiles cleanly.

Infrastructure: pyproject.toml, Dockerfile, Makefile, GitHub Actions CI.

+++ experiments/wrds_data_loader.py
@@ -128,7 +128,7 @@ def fetch_wrds_equities(
     )
     cached = _read_cache(cache_file)
     if cached is not None:
-        return cached
+        return cached.astype('float64')
 
     db = get_wrds_connection()
 
@@ -157,6 +157,8 @@ def fetch_wrds_equities(
     prices = df.pivot_table(index='date', columns='symbol', values='close')
     prices = prices.dropna()
     prices.index.name = None
+    # Ensure plain numpy float64 (parquet may restore pandas nullable Float64)
+    prices = prices.astype('float64')
 
     _write_cache(prices, cache_file)
     return prices
